requirements.txt
phidata
python-dotenv
fastapi
groq
duckduckgo-search
packaging
yfinance
uvicorn